requirements.txt
selenium
beautifulsoup4
python-dotenv
psycopg2-binary
tabulate
requests